Update README.md
These changes listed under readme-edits, which is a branch to the master repository, were made in accordance with the instructions given in the on-line tutorial for GitHub. Each instruction was precisely followed, to  a t. After creating a repository called hello-world, I proceeded to create a branch, distinct from the Master branch, which I named readme-edits. After making changes to this branch, I am making a comment on the change prior to committing it.